🏗️ Move components related to user feature into new dedicated folder
