+++ src/features/user/components/AvatarForm/style.scss
