+++ src/features/user/components/ProfileForm/index.jsx
@@ -4,12 +4,12 @@ import { useDispatch, useSelector } from 'react-redux'
 import { useForm } from 'react-hook-form'
 import { useLocation, useNavigate, useSearchParams } from 'react-router-dom'
 import { useEffect, useState } from 'react'
-import { getUser, getIsLogged } from '../../../redux/selectors/user'
-import { createUser, updateUser } from '../../../redux/reducers/user'
+import { getUser, getIsLogged } from '../../../../redux/selectors/user'
+import { createUser, updateUser } from '../../../../redux/reducers/user'
 import { AvatarForm } from '../AvatarForm'
-import { api, fetchCsrfCookie } from '../../../services/api'
-import { useServerErrors } from '../../../hooks/useServerErrors'
-import { ErrorCode, setErrorPage } from '../../../redux/reducers/errorPage'
+import { api, fetchCsrfCookie } from '../../../../services/api'
+import { useServerErrors } from '../../../../hooks/useServerErrors'
+import { ErrorCode, setErrorPage } from '../../../../redux/reducers/errorPage'
 
 import './style.scss'
